feat: read name action data
